refactor(PAT-03): split analyze.ts god module into domain modules for testability
Extract 5 new modules from the 2606-line analyze.ts:
- types.ts: all internal GraphQL response interfaces (DocBlob, RepoOverviewResponse, etc.)
- analyzer-utils.ts: pure utility functions (computeMedian, filterNodes*, getDurationHours, etc.)
- extract-responsiveness.ts: two-pass responsiveness fetch + metrics builders
- extract-contributors.ts: commit history pagination + contributor metrics builders
- extract-activity.ts: activity counts, cadence, discussion pagination, release health

analyze.ts is now a thin orchestrator that re-exports toAnalyzerError for backward compatibility.
Adds 77 new tests directly exercising previously-untestable internal functions.
